fix(#3): 어드민 검증 로직 변경 - CheckADminPort를 사용하지않고, ChatRoom 도메인에 검증 메서드를 사용하는 방식으로 변경했어요. - ChatRoomDomainErrorProperty를 구현했어요. - ChatRoom에서 어드민 아이디를 사용하는 모든 API에 유저 검증 로직을 추가했어요.
