--- src/main/java/com/teka/application/chatroom/exception/OverOperatingPeriodException.java
@@ -1,11 +1,0 @@
-package com.teka.application.chatroom.exception;
-
-import com.teka.application.chatroom.exception.error.ChatRoomApplicationErrorProperty;
-import com.teka.shared.error.TekaException;
-
-public class OverOperatingPeriodException extends TekaException {
-
-    public OverOperatingPeriodException() {
-        super(ChatRoomApplicationErrorProperty.OVER_OPERATING_PERIOD);
-    }
-}
