+++ src/main/java/com/teka/application/chatroom/service/ChangeUserTypeService.java
@@ -1,9 +1,9 @@
 package com.teka.application.chatroom.service;
 
-import com.teka.application.chatroom.exception.ChatRoomAdminMismatchException;
+import com.teka.application.chatroom.exception.ChatRoomNotFoundException;
 import com.teka.application.chatroom.port.in.ChangeUserTypeUseCase;
 import com.teka.application.chatroom.port.in.command.ChangeUserTypeCommand;
-import com.teka.application.chatroom.port.out.CheckAdminPort;
+import com.teka.application.chatroom.port.out.FindChatRoomPort;
 import com.teka.application.user.port.out.ChangeUserPort;
 import com.teka.domain.admin.AdminId;
 import com.teka.domain.chatroom.ChatRoomId;
@@ -17,13 +17,13 @@ import java.util.List;
 public class ChangeUserTypeService implements ChangeUserTypeUseCase {
 
     private final ChangeUserPort changeUserPort;
-    private final CheckAdminPort checkAdminPort;
+    private final FindChatRoomPort findChatRoomPort;
 
     @Override
     public void execute(List<ChangeUserTypeCommand> commandList, ChatRoomId chatRoomId, AdminId adminId) {
-        if (!checkAdminPort.checkChatRoomByAdminId(chatRoomId, adminId)) {
-            throw new ChatRoomAdminMismatchException();
-        }
+        findChatRoomPort.findById(chatRoomId)
+                .orElseThrow(ChatRoomNotFoundException::new)
+                .isAdmin(adminId);
         commandList.forEach(command -> changeUserPort.changeUserType(command.userId(), command.type()));
     }
 }
