+++ src/main/java/com/teka/application/chatroom/service/DeleteChatRoomService.java
@@ -1,9 +1,9 @@
 package com.teka.application.chatroom.service;
 
-import com.teka.application.chatroom.exception.ChatRoomAdminMismatchException;
+import com.teka.application.chatroom.exception.ChatRoomNotFoundException;
 import com.teka.application.chatroom.port.in.DeleteChatRoomUseCase;
-import com.teka.application.chatroom.port.out.CheckAdminPort;
 import com.teka.application.chatroom.port.out.DeleteChatRoomPort;
+import com.teka.application.chatroom.port.out.FindChatRoomPort;
 import com.teka.domain.admin.AdminId;
 import com.teka.domain.chatroom.ChatRoomId;
 import lombok.RequiredArgsConstructor;
@@ -14,13 +14,13 @@ import org.springframework.stereotype.Service;
 public class DeleteChatRoomService implements DeleteChatRoomUseCase {
 
     private final DeleteChatRoomPort deleteChatRoomPort;
-    private final CheckAdminPort checkAdminPort;
+    private final FindChatRoomPort findChatRoomPort;
 
     @Override
     public void execute(ChatRoomId chatRoomId, AdminId adminId) {
-        if (!checkAdminPort.checkChatRoomByAdminId(chatRoomId, adminId)) {
-            throw new ChatRoomAdminMismatchException();
-        }
+        findChatRoomPort.findById(chatRoomId)
+                .orElseThrow(ChatRoomNotFoundException::new)
+                .isAdmin(adminId);
         deleteChatRoomPort.deleteById(chatRoomId);
     }
 }
