+++ src/main/java/com/teka/application/chatroom/service/DeleteUserService.java
@@ -1,9 +1,10 @@
 package com.teka.application.chatroom.service;
 
-import com.teka.application.chatroom.exception.ChatRoomAdminMismatchException;
+import com.teka.application.chatroom.exception.ChatRoomNotFoundException;
+import com.teka.application.chatroom.port.out.FindChatRoomPort;
+import com.teka.domain.chatroom.exception.ChatRoomAdminMismatchException;
 import com.teka.application.chatroom.port.in.DeleteUserUseCase;
 import com.teka.application.chatroom.port.in.command.DeleteUserCommand;
-import com.teka.application.chatroom.port.out.CheckAdminPort;
 import com.teka.application.user.port.out.DeleteUserPort;
 import com.teka.domain.admin.AdminId;
 import com.teka.domain.chatroom.ChatRoomId;
@@ -17,13 +18,13 @@ import java.util.List;
 public class DeleteUserService implements DeleteUserUseCase {
 
     private final DeleteUserPort deleteUserPort;
-    private final CheckAdminPort checkAdminPort;
+    private final FindChatRoomPort findChatRoomPort;
 
     @Override
     public void execute(List<DeleteUserCommand> commandList, ChatRoomId chatRoomId, AdminId adminId) {
-        if (!checkAdminPort.checkChatRoomByAdminId(chatRoomId, adminId)) {
-            throw new ChatRoomAdminMismatchException();
-        }
+        findChatRoomPort.findById(chatRoomId)
+                .orElseThrow(ChatRoomNotFoundException::new)
+                .isAdmin(adminId);
         commandList.forEach(command -> deleteUserPort.deleteByUserId(command.userId()));
     }
 }
