+++ src/main/java/com/teka/application/chatroom/service/OpenChatRoomService.java
@@ -1,10 +1,7 @@
 package com.teka.application.chatroom.service;
 
-import com.teka.application.chatroom.exception.ChatRoomAdminMismatchException;
 import com.teka.application.chatroom.exception.ChatRoomNotFoundException;
-import com.teka.application.chatroom.exception.OverOperatingPeriodException;
 import com.teka.application.chatroom.port.in.OpenChatRoomUseCase;
-import com.teka.application.chatroom.port.out.CheckAdminPort;
 import com.teka.application.chatroom.port.out.FindChatRoomPort;
 import com.teka.application.chatroom.port.out.OpenChatRoomPort;
 import com.teka.domain.admin.AdminId;
@@ -21,19 +18,13 @@ public class OpenChatRoomService implements OpenChatRoomUseCase {
 
     private final OpenChatRoomPort openChatRoomPort;
     private final FindChatRoomPort findChatRoomPort;
-    private final CheckAdminPort checkAdminPort;
 
     @Override
     public void execute(ChatRoomId chatRoomId, AdminId adminId) {
-        if (!checkAdminPort.checkChatRoomByAdminId(chatRoomId, adminId)) {
-            throw new ChatRoomAdminMismatchException();
-        }
         ChatRoom chatRoom = findChatRoomPort.findById(chatRoomId)
                 .orElseThrow(ChatRoomNotFoundException::new);
-        LocalDate now = LocalDate.now();
-        if (now.isAfter(chatRoom.getEndDate())) {
-            throw new OverOperatingPeriodException();
-        }
+        chatRoom.isAdmin(adminId);
+        chatRoom.validateOperatingPeriod(LocalDate.now());
 
         openChatRoomPort.open(chatRoomId);
     }
