+++ src/main/java/com/teka/application/chatroom/service/RegisterUserService.java
@@ -1,10 +1,10 @@
 package com.teka.application.chatroom.service;
 
-import com.teka.application.chatroom.exception.ChatRoomAdminMismatchException;
+import com.teka.application.chatroom.port.out.FindChatRoomPort;
+import com.teka.domain.chatroom.exception.ChatRoomAdminMismatchException;
 import com.teka.application.chatroom.exception.ChatRoomNotFoundException;
 import com.teka.application.chatroom.port.in.RegisterUserUseCase;
 import com.teka.application.chatroom.port.in.command.RegisterUserCommand;
-import com.teka.application.chatroom.port.out.CheckAdminPort;
 import com.teka.application.user.port.in.exception.EmailAlreadyExistsException;
 import com.teka.application.user.port.in.exception.PhoneNumberAlreadyExistsException;
 import com.teka.application.user.port.out.CheckUserEmailPort;
@@ -28,34 +28,29 @@ public class RegisterUserService implements RegisterUserUseCase {
     private final SaveUserPort saveUserPort;
     private final CheckUserPhoneNumberPort checkUserPhoneNumberPort;
     private final CheckUserEmailPort checkUserEmailPort;
-    private final CheckAdminPort checkAdminPort;
+    private final FindChatRoomPort findChatRoomPort;
 
     @Override
     public void execute(List<RegisterUserCommand> commandList, ChatRoomId chatRoomId, AdminId adminId) {
-        if(!checkAdminPort.checkChatRoomByAdminId(chatRoomId, adminId)) {
-            throw new ChatRoomAdminMismatchException();
-        }
+        findChatRoomPort.findById(chatRoomId)
+                .orElseThrow(ChatRoomNotFoundException::new)
+                .isAdmin(adminId);
         for (RegisterUserCommand command : commandList) {
-            try {
-                if (checkUserPhoneNumberPort.existsByPhoneNumber(command.phoneNumber(), chatRoomId)) {
-                    throw new PhoneNumberAlreadyExistsException();
-                }
-                if (checkUserEmailPort.existsByEmail(command.email(), chatRoomId)) {
-                    throw new EmailAlreadyExistsException();
-                }
-
-                User user = User.builder()
-                        .id(null)
-                        .phoneNumber(new PhoneNumber(command.phoneNumber()))
-                        .email(new Email(command.email()))
-                        .type(command.type())
-                        .chatRoomId(chatRoomId)
-                        .build();
-                saveUserPort.save(user);
-            } catch (EntityNotFoundException e) {
-                throw new ChatRoomNotFoundException();
-
+            if (checkUserPhoneNumberPort.existsByPhoneNumber(command.phoneNumber(), chatRoomId)) {
+                throw new PhoneNumberAlreadyExistsException();
             }
+            if (checkUserEmailPort.existsByEmail(command.email(), chatRoomId)) {
+                throw new EmailAlreadyExistsException();
+            }
+
+            User user = User.builder()
+                    .id(null)
+                    .phoneNumber(new PhoneNumber(command.phoneNumber()))
+                    .email(new Email(command.email()))
+                    .type(command.type())
+                    .chatRoomId(chatRoomId)
+                    .build();
+            saveUserPort.save(user);
         }
     }
 }
